phase drawing fix

+++ DynModels/resultwidget.h
@@ -34,7 +34,7 @@ private:
     QList<ASolveByMethod>   results;
     QStringList roles;
     void checkClicked(const QCheckBox *cb, const DiffMethod &method, const quint8 &index);
-    quint8 getSizeOfData() const;
+    int  getSizeOfData() const;
     void draw2DPlots(const quint8 &index, const DiffMethod &method);
 signals:
     void sigMouseHoveredOn2DPlots(const QPointF &point);
